feat: :sparkles: enhance caching mechanism with IS_CACHE_ENABLED flag
- Introduced IS_CACHE_ENABLED environment variable to control caching behavior across various API routes.
- Updated multiple files to replace SUBSCAN_CACHE_ENABLED with IS_CACHE_ENABLED, ensuring consistent cache management.
- Modified caching logic in fetchSubscanData, route handlers, and comment/reaction management to respect the new caching flag.
- Improved overall code maintainability and flexibility regarding caching strategies.

+++ src/app/api/v2/[proposalType]/route.ts
@@ -20,6 +20,7 @@ import { AuthService } from '../../_api-services/auth_service';
 import { getReqBody } from '../../_api-utils/getReqBody';
 import { convertContentForFirestoreServer } from '../../_api-utils/convertContentForFirestoreServer';
 import { RedisService } from '../../_api-services/redis_service';
+import { IS_CACHE_ENABLED } from '../../_api-constants/apiEnvVars';
 
 const zodParamsSchema = z.object({
 	proposalType: z.nativeEnum(EProposalType)
@@ -44,7 +45,7 @@ export const GET = withErrorHandling(async (req: NextRequest, { params }) => {
 
 	// Try to get from cache first
 	const cachedData = await RedisService.GetPostsListing({ network, proposalType, page, limit, statuses, origins, tags });
-	if (cachedData) {
+	if (cachedData && IS_CACHE_ENABLED) {
 		return NextResponse.json(deepParseJson(cachedData));
 	}
 
@@ -141,7 +142,9 @@ export const GET = withErrorHandling(async (req: NextRequest, { params }) => {
 	};
 
 	// Cache the response
-	await RedisService.SetPostsListing({ network, proposalType, page, limit, data: JSON.stringify(response), statuses, origins, tags });
+	if (IS_CACHE_ENABLED) {
+		await RedisService.SetPostsListing({ network, proposalType, page, limit, data: JSON.stringify(response), statuses, origins, tags });
+	}
 
 	// 3. return the data
 	return NextResponse.json(response);
@@ -177,8 +180,10 @@ export const POST = withErrorHandling(async (req: NextRequest, { params }: { par
 	});
 
 	// Invalidate post listings since a new post was added
-	await RedisService.DeletePostsListing({ network, proposalType });
-	await RedisService.DeleteActivityFeed({ network }); // Invalidate activity feed since a new post was added
+	if (IS_CACHE_ENABLED) {
+		await RedisService.DeletePostsListing({ network, proposalType });
+		await RedisService.DeleteActivityFeed({ network }); // Invalidate activity feed since a new post was added
+	}
 
 	const response = NextResponse.json({ message: 'Post created successfully', data: { id, index: Number(indexOrHash) } });
 	response.headers.append('Set-Cookie', await AuthService.GetAccessTokenCookie(newAccessToken));
